requirements.txt
pyyaml>=6.0
python-dotenv>=1.0.0
requests>=2.28.0
openai>=1.0.0
playwright>=1.40.0
pypdf>=3.0.0
streamlit>=1.30.0
